+++ CHANGELOG.md
@@ -7,7 +7,8 @@ A new section Unreleased is opened then for next dev phase.
 ## Unreleased
 
 ### Added
-- Add notebooks for demcompare users 
+- Add coregistration notebook for demcompare users 
+- Add statistics notebook for demcompare users 
 
 ### Changed
 - Refactoring image and dem tools with demcompare dataset. [#76]